# AVX512 accelerated ED25519 and X25519 implementations