… src/ballet/ed25519/fd_ed25519_user.c
… src/ballet/ed25519/fd_x25519.c
… src/ballet/ed25519/table/fd_r43x6_ge_dmul_sparse_table
… src/ballet/ed25519/table/fd_r43x6_ge_smul_base_large_table
… src/ballet/ed25519/table/fd_r43x6_ge_smul_base_small_table
… src/util/fd_util.h
… src/util/fd_util_base.h
… src/util/simd/Local.mk
… src/util/simd/fd_avx512.h
… src/util/simd/fd_avx512_wwl.h
… src/util/simd/fd_avx_wd.h
… src/util/simd/fd_avx_wf.h
… src/util/simd/fd_sse_vd.h
… src/util/simd/fd_sse_vf.h
… src/util/simd/test_avx512_8x64.c
… src/util/simd/test_avx_4x64.c
… src/util/simd/test_avx_8x32.c
… src/util/simd/test_sse_16x8.c
… src/util/simd/test_sse_2x64.c
… src/util/simd/test_sse_4x32.c
… src/util/simd/test_sse_common.c
… src/util/test_util_base.c