This replaces PR #647

Quick and dirty benchmarks on a single icelake-server core at 2.3 GHz under gcc-12 yields:

```
       |         implementation         |    speed up over
op     | portable |     AVX2 |   AVX512 | portable |     AVX2
-------+----------+----------+----------+----------+---------
x25519 | 135.0 us |  84.8 us |  34.7 us |     3.9x |     2.4x
keygen |  46.4 us |  25.0 us |  11.4 us |     4.1x |     2.2x
sign   |  49.5 us |  26.8 us |  13.2 us |     3.8x |     2.0x
verify | 154.4 us |  70.3 us |  34.2 us |     4.5x |     2.1x
```

Roughly, this implementation is ~4x faster than the existing scalar implementation and ~2x faster than the existing AVX2 implementation on the same AVX512 supporting hardware with the same compiler and build flags.  YMMV for other processor architectures.  Wallclock performance is expected for operations like this to scale with clock speed.

(clang-15 does not seem to optimize AVX512 as well as gcc-12 and runs the AVX512 implementation 10-15% slower.)

Details documented extensively in the code and commit comments, including additional potential algorithmic and implementation improvements, areas in which other implementations diverge from the standard and how to tweak this implementation to match an particular implementation behaviors.