@kbowers-jump 2023-09-23
Since Marc was asking, the above measurements was with Turbo mode disabled (specifically the "turbo mode disable bit", bit 38 in MSR 0x1a0 was set ... for reference, typically run benchmarks without turbo mode to get stable results, especially when benchmarking an _implementation_ ... real pain to do, say, A/B testing to tune up an implementation when Intel is changing the clock speed between and during runs on a whim).

The exact processor is a Platform 8380: https://www.intel.com/content/www/us/en/products/sku/212287/intel-xeon-platinum-8380-processor-60m-cache-2-30-ghz/specifications.html

Did a quick and dirty experiment on the same core with Turbo mode enabled (but didn't poke around all the clock frequency ratio / thermal envelope / voltage MSRs to try to max clock frequency or stabilize system performance ... Intel tends to not let AVX-512 heavy code hit the top frequencies anyway so have no clue how close or stable the clock got to Turbo max speed of 3.4 GHz).  This got X25519 to 27.5 us, keygen at 9.1 us, signing at 10.5 us, verify at 27.7 us ... about +25% faster.  YMMV.