@kaasaraai-jump 2023-09-27
Javier, Emanuele and I have been going through the code, looks impressive / nice-to-follow so far.  Specifically so far we've covered ED25519 mul-modp at bit-level and sig-verify / signing at data-flow level.